backend/requirements.txt
anyio==3.6.1
ariadne==0.16.1
async-timeout==4.0.2
authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
confuse==2.0.0
cryptography==38.0.1
deprecated==1.2.13
graphql-core==3.2.3
greenlet==1.1.3.post0
h11==0.14.0
idna==3.4
packaging==21.3
pandas==2.2.3
plyse==1.0.3
psycopg2-binary==2.9.4
pycparser==2.21
pydantic==2.4.2
pydantic-core==2.10.1
pyparsing==2.4.0
pyyaml==6.0
redis==4.3.5
requests==2.28.1
sniffio==1.3.0
sqlalchemy==2.0.20
starlette==0.20.4
starlette-context==0.3.6
typing-extensions==4.7.1
urllib3==1.26.13
uvicorn==0.18.3
wrapt==1.14.1
